Main loop works now
Refactor message related stuff into messages.rs

handle.add() is unsafe because it requires that the handle is not moved
after add is called. The changes in chatbot.run are to account for this.

The main loop simply echoes the payload back to where it came from for
now. Next step is to iterate through any attached handlers.