Add GithubIssueLinker handler
Messages arriving that include a github issue URL will have their title
and URL included in the reply. As the first non-trivial handler, this is
more verbose than need be.

The MessageHandler trait should be updated to require a `.re` method
which returns a regular expression that tests if the handler is
interested in the message. Some default implementations for helper
functions like `can_handle` and `get_capturing_groups` can be supplied.
The `handle` method could be updated to accept an Option for Capturing.

There's a lot of `unwrap` calls in this handler that make it dangerous.
Before writing many more handlers, this needs to be solved.

The CliAdapter was updated to flush when receiving an outgoing message.