Remove remaining unstable API usage
The select! macro turned out to be unstable as well (not clear in docs
but very clear in source). It turned out not to be necessary in the CLI
adapter anyway, and that code is a lot cleaner now.

regex! usage is still available although they are dynamic instead of
native code.

The Cargo.toml file was updated for a release.