Add ping/pong handler
Incoming messages containing "ping" will be responded to with "pong".
It's becoming important that some handlers can opt into direct
addressing. I think it's safe to say most people would not appreciate
the bot interjecting every time somebody says "ping".

The ping handler being 68 lines of code is quite the bummer. 29 of those
lines are tests, and most of the remaining lines are boiler plate for
integrating with the chatbot. There are basically 2 lines that dictate
the behavior of the handler. In short, it is time for a refactor.