Add support (again) for concurrent adapters
The original implementation of concurrent adapters relied on
std::sync::mpsc::Select which is unstable and unsafe. Since chatbot
needed to be released on the stable line, using Select was not an
option. The adapter interface was refactored slightly to accept a sender
instead of returning a receiver. The channel is now created by the main
thread, and transmitters for IncomingMessage are included in the adapter
`process_events` call. This allows the receiver in the main thread to
receive messages from all of the adapters.

Resolves #13.