Start IRC adapter implemntation
The IrcAdapter is properly receiving PRIVMSG commands and transforming
them into chatbot::message::IncomingMessage to be consumed elsewhere in
the bot. The outgoing portion of the adapter still needs to be
implemented.

The ChatAdapter trait was updated to include a Chatbot reference in the
process_events arguments. This is necessary for `reply` handling to
route the message correctly (private vs channel).

This will resolve #1 once the rest of the adapter is implemented.