Add IrcAdapter outgoing message support
The IrcAdapter can now receive OutgoingMessages and send them to the
channel or user where the original message came from. Right now it is
assumed that the IRC nick is the same as the bot's name for determining
whether to reply to the user or channel. It would be nice if the `irc`
crate had a way for determining which nick it was able to connect with.

The Chatbot::new method now takes a name string instead of being
hardcoded to "computer". The README was even updated to reflect that
change (yay).